requirements.txt
transformers
torch
sentence-transformers
faiss-cpu
PyPDF2
numpy
tqdm
